slidebar
This file creates a sidebar section with the ID "sidebar". It then checks if the sidebar-1 widget area is active, and if so, it displays the widgets that have been added to that area using the *dynamic_sidebar function. You can customize the appearance and functionality of your sidebar widgets using the WordPress Widget API.